fixed contact infor display
