+++ src/pages/contact/index.js
@@ -1,5 +1,5 @@
 import contactForm from "./contactForm";
-import { email, phone } from "../footer/contactDetails";
+import { email, phone } from "./contactDetails";
 
 const contact = document.createElement("section");
 contact.setAttribute("id", "contact");
@@ -8,8 +8,8 @@ header.textContent = "Contact";
 const paragraph = document.createElement("p");
 paragraph.textContent = "Looking forward to hearing from you.";
 const contactsContainer = document.createElement("div");
-contactsContainer.appendChild(phone);
-contactsContainer.appendChild(email);
+contactsContainer.appendChild(phone());
+contactsContainer.appendChild(email());
 const div = document.createElement("div");
 div.appendChild(contactsContainer);
 div.appendChild(contactForm);
